refactor(home): rename header file to home_header
